+++ test/screens/home/views/refuelings_test.dart
@@ -19,14 +19,18 @@ class MockFilteredRefuelingsBloc
     extends MockBloc<FilteredRefuelingsLoaded, FilteredRefuelingsState>
     implements FilteredRefuelingsBloc {}
 
+class MockCarsBloc extends MockBloc<CarsEvent, CarsState> implements CarsBloc {}
+
 void main() {
   RefuelingsBloc refuelingsBloc;
   FilteredRefuelingsBloc filteredRefuelingsBloc;
+  CarsBloc carsBloc;
   BasePrefService pref;
 
   setUp(() async {
     refuelingsBloc = MockRefuelingsBloc();
     filteredRefuelingsBloc = MockFilteredRefuelingsBloc();
+    carsBloc = MockCarsBloc();
     pref = JustCachePrefService();
     await pref.setDefaultValues({
       'length_unit': DistanceUnit.imperial.index,
@@ -36,33 +40,6 @@ void main() {
   });
 
   group('RefuelingsScreen', () {
-    testWidgets('loading', (WidgetTester tester) async {
-      when(filteredRefuelingsBloc.state)
-          .thenAnswer((_) => FilteredRefuelingsLoading());
-      final refuelingsKey = Key('refuelings');
-      await tester.pumpWidget(
-        ChangeNotifierProvider<BasePrefService>.value(
-          value: pref,
-          child: MultiBlocProvider(
-            providers: [
-              BlocProvider<RefuelingsBloc>.value(
-                value: refuelingsBloc,
-              ),
-              BlocProvider<FilteredRefuelingsBloc>.value(
-                value: filteredRefuelingsBloc,
-              ),
-            ],
-            child: MaterialApp(
-              home: Scaffold(
-                body: RefuelingsScreen(key: refuelingsKey),
-              ),
-            ),
-          ),
-        ),
-      );
-      await tester.pump();
-      expect(find.byType(LoadingIndicator), findsOneWidget);
-    });
     testWidgets('not loaded', (WidgetTester tester) async {
       when(filteredRefuelingsBloc.state)
           .thenAnswer((_) => FilteredRefuelingsNotLoaded());
@@ -76,6 +53,7 @@ void main() {
             BlocProvider<FilteredRefuelingsBloc>.value(
               value: filteredRefuelingsBloc,
             ),
+            BlocProvider<CarsBloc>.value(value: carsBloc),
           ],
           child: MaterialApp(
             home: Scaffold(
@@ -87,7 +65,7 @@ void main() {
       await tester.pump();
       expect(find.byKey(refuelingsKey), findsOneWidget);
     });
-    testWidgets('tap', (WidgetTester tester) async {
+    testWidgets('load', (WidgetTester tester) async {
       final refueling = Refueling(
         carName: 'test',
         amount: 10.0,
@@ -96,7 +74,8 @@ void main() {
         date: DateTime.fromMillisecondsSinceEpoch(0),
       );
       when(filteredRefuelingsBloc.state).thenAnswer((_) =>
-          FilteredRefuelingsLoaded([refueling], VisibilityFilter.all, [Car()]));
+          FilteredRefuelingsLoaded([refueling], VisibilityFilter.all, [Car(name: 'test')]));
+      when(carsBloc.state).thenReturn(CarsLoaded([Car(name: 'test')]));
       final refuelingsKey = Key('refuelings');
       await tester.pumpWidget(
         ChangeNotifierProvider<BasePrefService>.value(
@@ -109,6 +88,7 @@ void main() {
               BlocProvider<FilteredRefuelingsBloc>.value(
                 value: filteredRefuelingsBloc,
               ),
+              BlocProvider<CarsBloc>.value(value: carsBloc),
             ],
             child: MaterialApp(
               home: Scaffold(
@@ -119,81 +99,7 @@ void main() {
         ),
       );
       await tester.pumpAndSettle();
-      await tester.tap(find.byType(RefuelingCard));
-      await tester.pump();
-      expect(find.byKey(refuelingsKey), findsOneWidget);
-    });
-    testWidgets('dismiss', (WidgetTester tester) async {
-      final refueling = Refueling(
-        carName: 'test',
-        amount: 10.0,
-        cost: 10.0,
-        mileage: 1000,
-        date: DateTime.fromMillisecondsSinceEpoch(0),
-      );
-      when(filteredRefuelingsBloc.state).thenAnswer((_) =>
-          FilteredRefuelingsLoaded([refueling], VisibilityFilter.all, [Car()]));
-      final refuelingsKey = Key('refuelings');
-      await tester.pumpWidget(
-        ChangeNotifierProvider<BasePrefService>.value(
-          value: pref,
-          child: MultiBlocProvider(
-            providers: [
-              BlocProvider<RefuelingsBloc>.value(
-                value: refuelingsBloc,
-              ),
-              BlocProvider<FilteredRefuelingsBloc>.value(
-                value: filteredRefuelingsBloc,
-              ),
-            ],
-            child: MaterialApp(
-              home: Scaffold(
-                body: RefuelingsScreen(key: refuelingsKey),
-              ),
-            ),
-          ),
-        ),
-      );
-      await tester.pumpAndSettle();
-      await tester.fling(find.byType(RefuelingCard), Offset(-300, 0), 10000.0);
-      await tester.pumpAndSettle();
-      verify(refuelingsBloc.add(DeleteRefueling(refueling))).called(1);
-    });
-    testWidgets('tap', (WidgetTester tester) async {
-      final refueling = Refueling(
-        carName: 'test',
-        amount: 10.0,
-        cost: 10.0,
-        mileage: 1000,
-        date: DateTime.fromMillisecondsSinceEpoch(0),
-      );
-      when(filteredRefuelingsBloc.state).thenAnswer((_) =>
-          FilteredRefuelingsLoaded([refueling], VisibilityFilter.all, [Car()]));
-      final refuelingsKey = Key('refuelings');
-      await tester.pumpWidget(
-        ChangeNotifierProvider<BasePrefService>.value(
-          value: pref,
-          child: MultiBlocProvider(
-            providers: [
-              BlocProvider<RefuelingsBloc>.value(
-                value: refuelingsBloc,
-              ),
-              BlocProvider<FilteredRefuelingsBloc>.value(
-                value: filteredRefuelingsBloc,
-              ),
-            ],
-            child: MaterialApp(
-              home: Scaffold(
-                body: RefuelingsScreen(key: refuelingsKey),
-              ),
-            ),
-          ),
-        ),
-      );
-      await tester.pumpAndSettle();
-      await tester.tap(find.byType(RefuelingCard));
-      await tester.pumpAndSettle();
-      expect(find.byKey(refuelingsKey), findsOneWidget);
+      expect(find.byType(RefuelingCard), findsOneWidget);
     });
   });
 }
